fxa-client: Remove BrowserID assertions, and their feature flag.
This removes support for generating BrowserID assertions, and moves some
of the other session-token-using machinery that was beind the "browserid"
into the main package, where it is now enabled by default.
