+++ components/fxa-client/ffi/Cargo.toml
@@ -21,7 +21,6 @@ viaduct = { path = "../../viaduct" }
 path = "../"
 
 [features]
-browserid = ["fxa-client/browserid"]
 reqwest = ["viaduct/reqwest", "fxa-client/reqwest"]
 
 [target.'cfg(target_os = "android")'.dependencies]
