+++ components/fxa-client/ffi/src/lib.rs
@@ -247,7 +247,7 @@ pub extern "C" fn fxa_complete_oauth_flow(
     });
 }
 
-/// Migrate from a logged-in browserid Firefox Account.
+/// Migrate from a logged-in sessionToken Firefox Account.
 #[no_mangle]
 pub unsafe extern "C" fn fxa_migrate_from_session_token(
     handle: u64,
