+++ components/fxa-client/src/state_persistence.rs
@@ -81,8 +81,6 @@ impl From<StateV1> for Result<StateV2> {
                 state.client_id,
                 state.redirect_uri,
             ),
-            #[cfg(feature = "browserid")]
-            login_state: super::login_sm::LoginState::Unknown,
             refresh_token,
             scoped_keys: all_scoped_keys,
             last_handled_command: None,
@@ -99,8 +97,6 @@ struct StateV1 {
     client_id: String,
     redirect_uri: String,
     config: V1Config,
-    // #[cfg(feature = "browserid")]
-    // login_state: LoginState, // Wasn't used anyway
     oauth_cache: HashMap<String, V1AuthInfo>,
 }
